# docs/swagger.yaml
definitions:
...
  models.Projects:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      year:
        type: string
    type: object
...
  response.BaseResponse-models_Projects:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.Projects'
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /projects/{id}:
    get:
      consumes:
      - application/json
      description: Get specific project data by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse-models_Projects'
      summary: Get Project Detail
      tags:
      - Projects
swagger: "2.0"